Added a connection popover for the output side, fixed boundary condition bugs when dragging connections.

+++ WMEdit/WMEdit/Graph View/Patch View/WMPatchPlugStripView.m
@@ -74,7 +74,7 @@
 	if (inputCount > 0) {
 		CGFloat offX = (inPoint.x - leftOffset) / offsetBetweenDots;
 		int offXi = (int)roundf(offX);
-		return MAX(0, MIN(offXi, inputCount - 1));
+		return MAX(0, MIN(offXi, (int)inputCount - 1));
 	}
 	return NSNotFound;
 }
